Created strikeZone class to detect both a horiz and vertical LDR at same
This is the first version that tries to detect a strike by looking for
a horizontal and vertical trigger at the same time. The problem with
this version is you will always get a ball first time as it immediately
checks for a strike. In this case there will never be both.