Two LDR sesnors now working in all cases
Fixed the issue with the first sensor triggering a ball or strike. We
added logic to wait for a second trigger. If the second trigger is the
other sensor then strike if its the first sensor going back to light
then ball. Also added logic so that if a strike is detected, we wait for
both the sensors to go back to light before waiting for a new trigger.